Classification Error

Measure to compare true observed labels with predicted labels in multiclass classification tasks.

ce(truth, response, sample_weights = NULL, ...)

Arguments

truth

(factor())
True (observed) labels. Must have the same levels and length as response.

response

(factor())
Predicted response labels. Must have the same levels and length as truth.

sample_weights

(numeric())
Vector of non-negative and finite sample weights. Must have the same length as truth. The vector gets automatically normalized to sum to one. Defaults to equal sample weights.

...

(any)
Additional arguments. Currently ignored.

Value

Performance value as numeric(1).

Details

The Classification Error is defined as $$ \frac{1}{n} \sum_{i=1}^n w_i \left( t_i \neq r_i \right). $$

Meta Information

Type: "classif"

Range: \([0, 1]\)

Minimize: TRUE

Required prediction: response

See also

Examples

set.seed(1)
lvls = c("a", "b", "c")
truth = factor(sample(lvls, 10, replace = TRUE), levels = lvls)
response = factor(sample(lvls, 10, replace = TRUE), levels = lvls)
ce(truth, response)
#> [1] 0.8