Symmetric Mean Absolute Percent Error

Measure to compare true observed response with predicted response in regression tasks.

smape(truth, response, na_value = NaN, ...)

Arguments

truth

(numeric())
True (observed) values. Must have the same length as response.

response

(numeric())
Predicted response values. Must have the same length as truth.

na_value

(numeric(1))
Value that should be returned if the measure is not defined for the input (as described in the note). Default is NaN.

...

(any)
Additional arguments. Currently ignored.

Value

Performance value as numeric(1).

Details

The Symmetric Mean Absolute Percent Error is defined as $$ \frac{2}{n} \sum_{i=1}^n \frac{\left| t_i - r_i \right|}{\left| t_i \right| + \left| r_i \right|}. $$

This measure is undefined if if any \(|t| + |r|\) is \(0\).

Meta Information

Type: "regr"

Range: \([0, 2]\)

Minimize: TRUE

Required prediction: response

See also

Examples

set.seed(1)
truth = 1:10
response = truth + rnorm(10)
smape(truth, response)
#> [1] 0.2115182