True Negatives

Measure to compare true observed labels with predicted labels in binary classification tasks.

tn(truth, response, positive, ...)

Arguments

truth

(factor())
True (observed) labels. Must have the exactly same two levels and the same length as response.

response

(factor())
Predicted response labels. Must have the exactly same two levels and the same length as truth.

positive

(character(1))
Name of the positive class.

...

(any)
Additional arguments. Currently ignored.

Value

Performance value as numeric(1).

Details

This measure counts the true negatives, i.e. the number of predictions correctly indicating a negative class label.

Meta Information

Type: "binary"

Range: \([0, \infty)\)

Minimize: FALSE

Required prediction: response

References

https://en.wikipedia.org/wiki/Template:DiagnosticTesting_Diagram

See also

Examples

set.seed(1)
lvls = c("a", "b")
truth = factor(sample(lvls, 10, replace = TRUE), levels = lvls)
response = factor(sample(lvls, 10, replace = TRUE), levels = lvls)
tn(truth, response, positive = "a")
#> [1] 1