Measure Registry

The environment() measures keeps track of all measures in this package. It stores meta information such as minimum, maximum or if the measure must be minimized or maximized. The following information is available for each measure:

id: Name of the measure.

title: Short descriptive title.

type: "binary" for binary classification, "classif" for binary or multi-class classification, "regr" for regression and "similarity" for similarity measures.

lower: lower bound.

upper: upper bound.

predict_type: prediction type the measure operates on. "response" corresponds to class labels for classification and the numeric response for regression. "prob" corresponds to class probabilities, provided as a matrix with class labels as column names. "se" corresponds to to the vector of predicted standard errors for regression.

minimize: If TRUE or FALSE, the objective is to minimize or maximize the measure, respectively. Can also be NA.

sample_weights: If TRUE, it is possible calculate a weighted measure.

measures

Format

An object of class environment of length 54.

Examples

names(measures)
#>  [1] "tp"          "mauc_au1p"   "fbeta"       "fnr"         "mbrier"     
#>  [6] "bacc"        "jaccard"     "ktau"        "maxse"       "mape"       
#> [11] "maxae"       "mauc_au1u"   "ce"          "ppv"         "sensitivity"
#> [16] "specificity" "phi"         "recall"      "fn"          "precision"  
#> [21] "srho"        "fp"          "tpr"         "bbrier"      "bias"       
#> [26] "acc"         "mcc"         "auc"         "rae"         "rrse"       
#> [31] "smape"       "rse"         "npv"         "mauc_aunp"   "rsq"        
#> [36] "tnr"         "fomr"        "fdr"         "fpr"         "dor"        
#> [41] "pbias"       "logloss"     "prauc"       "mae"         "mauc_aunu"  
#> [46] "rmsle"       "rmse"        "medse"       "sse"         "medae"      
#> [51] "msle"        "sae"         "mse"         "tn"         
measures$tpr
#> $id
#> [1] "tpr"
#> 
#> $title
#> [1] "True Positive Rate"
#> 
#> $type
#> [1] "binary"
#> 
#> $lower
#> [1] 0
#> 
#> $upper
#> [1] 1
#> 
#> $predict_type
#> [1] "response"
#> 
#> $minimize
#> [1] FALSE
#> 
#> $sample_weights
#> [1] FALSE
#>