Measure Registry

The environment() measures keeps track of all measures in this package. Stores meta information about measures, such as minimum, maximum of if the measure must be minimized or maximized.

measures

Format

An object of class environment of length 47.

Examples

names(measures)

#>  [1] "tpr"         "bacc"        "auc"         "mse"         "ce"         
#>  [6] "rse"         "tn"          "precision"   "tp"          "specificity"
#> [11] "sensitivity" "fbeta"       "tnr"         "dor"         "bbrier"     
#> [16] "recall"      "sae"         "msle"        "medae"       "rsq"        
#> [21] "fpr"         "sse"         "medse"       "ppv"         "fn"         
#> [26] "fp"          "fdr"         "fomr"        "mbrier"      "fnr"        
#> [31] "mcc"         "logloss"     "pbias"       "smape"       "maxae"      
#> [36] "rmse"        "mape"        "rrse"        "npv"         "acc"        
#> [41] "bias"        "maxse"       "rae"         "ktau"        "mae"        
#> [46] "rmsle"       "srho"       

measures$tpr

#> $id
#> [1] "tpr"
#> 
#> $type
#> [1] "binary"
#> 
#> $title
#> [1] "True Positive Rate"
#> 
#> $lower
#> [1] 0
#> 
#> $upper
#> [1] 1
#> 
#> $predict_type
#> [1] "response"
#> 
#> $minimize
#> [1] FALSE
#>