Relative Squared Error

Measure to compare true observed response with predicted response in regression tasks.

Usage

rse(truth, response, na_value = NaN, ...)

Arguments

truth

(numeric())
True (observed) values. Must have the same length as response.

response

(numeric())
Predicted response values. Must have the same length as truth.

na_value

(numeric(1))
Value that should be returned if the measure is not defined for the input (as described in the note). Default is NaN.

...

(any)
Additional arguments. Currently ignored.

Value

Performance value as numeric(1).

Details

The Relative Squared Error is defined as $$ \frac{\sum_{i=1}^n \left( t_i - r_i \right)^2}{\sum_{i=1}^n \left( t_i - \bar{t} \right)^2}. $$ Can be interpreted as squared error of the predictions relative to a naive model predicting the mean.

This measure is undefined for constant \(t\).

Meta Information

Type: "regr"

Range: \([0, \infty)\)

Minimize: TRUE

Required prediction: response

See also

Examples

set.seed(1)
truth = 1:10
response = truth + rnorm(10)
rse(truth, response)
#> [1] 0.06858916